requirements.txt
fastapi
langchain
langchain-community
langchain-openai
openai
faiss-cpu
python-dotenv
uvicorn
pydantic
pytest
httpx